requirements.txt
# Core
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
pyyaml==6.0.2
tenacity==9.0.0
httpx==0.27.2

# Data
pandas==2.2.3
numpy==1.26.4
pyarrow==17.0.0
polars==1.12.0

# Spark + Delta
pyspark==3.5.3
delta-spark==3.2.1

# ENTSO-E client
entsoe-py==0.6.11

# ML
scikit-learn==1.5.2
lightgbm==4.5.0
prophet==1.1.6
xgboost==2.1.1
shap==0.46.0

# Experiment tracking
mlflow==2.17.0

# Serving
fastapi==0.115.0
uvicorn[standard]==0.31.0

# Dashboard
streamlit==1.39.0
plotly==5.24.1
altair==5.4.1

# Monitoring
evidently==0.4.40
prometheus-client==0.21.0

# Observability
structlog==24.4.0
rich==13.9.2

# Dev
pytest==8.3.3
pytest-cov==5.0.0
pytest-mock==3.14.0
black==24.10.0
ruff==0.7.1
mypy==1.12.1
pre-commit==4.0.1
